refactor(parser): rename `models` modules to `domain` for consistent naming
Unify the dual `models`/`domain` naming convention across the workspace
under `domain`. Renames the three `models.rs` modules in `pubmed-parser`
(`common/models.rs`, `pubmed/models.rs`, `pmc/parser/models.rs`) and
updates all internal/downstream imports in `pubmed-client`,
`pubmed-formatter`, and `pubmed-cli` to use the new module path.
