+++ pubmed-parser/src/common/domain.rs
