+++ pubmed-parser/src/pmc/parser/mod.rs
@@ -3,8 +3,8 @@ use crate::error::Result;
 use crate::pmc::domain::PmcArticle;
 
 pub mod author;
+pub mod domain;
 pub mod metadata;
-pub mod models;
 pub(crate) mod reader_utils;
 pub mod reference;
 pub mod section;
