+++ pubmed-parser/src/pubmed/domain.rs
